+++ packages/studio/adapters/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-adapters@0.3.0...@opendaw/studio-adapters@0.3.1) (2026-08-24)
+
+**Note:** Version bump only for package @opendaw/studio-adapters
+
 # [0.3.0](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-adapters@0.2.1...@opendaw/studio-adapters@0.3.0) (2026-08-17)
 
 ### Features