+++ packages/studio/core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-core@0.2.2...@opendaw/studio-core@0.2.3) (2026-08-24)
+
+**Note:** Version bump only for package @opendaw/studio-core
+
 ## [0.2.2](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-core@0.2.1...@opendaw/studio-core@0.2.2) (2026-08-17)
 
 ### Bug Fixes